moved azure types into asset and installconfig
