+++ pkg/asset/installconfig/azure/session.go
