+++ pkg/asset/installconfig/platformcredscheck.go
@@ -6,6 +6,7 @@ import (
 	"github.com/gophercloud/utils/openstack/clientconfig"
 	"github.com/openshift/installer/pkg/asset"
 	awsconfig "github.com/openshift/installer/pkg/asset/installconfig/aws"
+	azureconfig "github.com/openshift/installer/pkg/asset/installconfig/aws"
 	"github.com/openshift/installer/pkg/types/aws"
 	"github.com/openshift/installer/pkg/types/azure"
 	"github.com/openshift/installer/pkg/types/libvirt"
@@ -53,7 +54,7 @@ func (a *PlatformCredsCheck) Generate(dependencies asset.Parents) error {
 	case libvirt.Name, none.Name, vsphere.Name:
 		// no creds to check
 	case azure.Name:
-		_, err = azure.GetSession()
+		_, err = azureconfig.GetSession()
 		if err != nil {
 			return errors.Wrap(err, "creating Azure session")
 		}
